Refactor FakeDatabase
